Move proj to ops
