+++ geopolars/src/lib.rs
@@ -5,9 +5,6 @@ pub mod ops;
 pub mod spatial_index;
 mod util;
 
-#[cfg(feature = "proj")]
-pub mod proj;
-
 #[cfg(test)]
 mod tests {
     #[test]
