+++ geopolars/src/ops/mod.rs
@@ -1,3 +1,5 @@
 pub mod affine;
 pub mod exterior;
 pub mod length;
+#[cfg(feature = "proj")]
+pub mod proj;
